requirements.txt
singer-python==5.12.2
requests==2.25.1
backoff==1.8.0